[Google Vision] Move "pdtToImages" from tesseract extractor to utils (will be reused by others extractors)
