+++ server/src/input/tesseract/TesseractExtractor.ts
@@ -14,11 +14,8 @@
  * limitations under the License.
  */
 
-import * as fs from 'fs';
-import * as path from 'path';
 import { Document } from '../../types/DocumentRepresentation';
-import * as utils from '../../utils';
-import logger from '../../utils/Logger';
+import { pdfToImages } from '../../utils';
 import { Extractor } from '../Extractor';
 import { setPageDimensions } from '../set-page-dimensions';
 import * as tesseract2json from './tesseract2json';
@@ -33,7 +30,7 @@ export class TesseractExtractor extends Extractor {
    * @returns The promise of a valid Document (as per the Document Representation namespace).
    */
   public async run(inputFile: string, rotationCorrection: boolean = true): Promise<Document> {
-    const imagePaths = await this.pdfToImages(inputFile);
+    const imagePaths = await pdfToImages(inputFile);
     return this.scanPages(imagePaths, rotationCorrection).then((doc: Document) => {
       doc.inputFile = inputFile;
       return doc;
@@ -60,48 +57,6 @@ export class TesseractExtractor extends Extractor {
       } else {
         return allPagesDoc;
       }
-    });
-  }
-
-  private pdfToImages(pdfPath: string): Promise<string[]> {
-    return new Promise((resolve, reject) => {
-      const folder = path.dirname(pdfPath).concat('/samples');
-      try {
-        if (!fs.existsSync(folder)) {
-          fs.mkdirSync(folder);
-        }
-      } catch (e) {
-        throw e;
-      }
-      const outPutFilePath = folder + '/Sample_%03d.tiff';
-      utils.CommandExecuter.run(utils.CommandExecuter.COMMANDS.CONVERT, [
-        pdfPath,
-        '-colorspace',
-        'RGB',
-        '-density',
-        '300x300',
-        '-compress',
-        'lzw',
-        '-alpha',
-        'remove',
-        '-background',
-        'white',
-        outPutFilePath,
-      ])
-        .then(() => {
-          const files = fs.readdirSync(folder).map(file => path.join(folder, file));
-          logger.info(`converted files: ${files.join(', ')}`);
-          resolve(files);
-        })
-        .catch(({ found, error }) => {
-          logger.error(error);
-          if (!found) {
-            logger.warn(
-              'ImageMagick failure: impossible to convert pdf to images (is ImageMagick installed?)',
-            );
-          }
-          reject(error);
-        });
     });
   }
 }
